+++ server/prisma/migrations/20260104043529_add_books_table/migration.sql
@@ -1,0 +1,14 @@
+-- CreateTable
+CREATE TABLE `Book` (
+    `id` INTEGER NOT NULL AUTO_INCREMENT,
+    `title` VARCHAR(200) NOT NULL,
+    `author` VARCHAR(150) NULL,
+    `imagePath` VARCHAR(300) NOT NULL,
+    `review` VARCHAR(500) NULL,
+    `createdAt` DATETIME(3) NOT NULL DEFAULT CURRENT_TIMESTAMP(3),
+    `updatedAt` DATETIME(3) NOT NULL,
+
+    INDEX `Book_title_idx`(`title`),
+    INDEX `Book_createdAt_idx`(`createdAt`),
+    PRIMARY KEY (`id`)
+) DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci;